phase-1 | refactor hooks
