+++ src/pages/Home/Home.page.jsx
@@ -1,27 +1,13 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { Alert, Spin } from 'antd';
-import { getVideos } from '../../utils/services/youtube';
 import VideoCard from '../../components/VideoCard/VideoCard.component';
 import { useSearch } from '../../providers/Search/Search.provider';
 import './Home.styles.css';
+import useFetchVideos from '../../utils/hooks/useFetchVideos';
 
 function HomePage() {
   const { searchTerm } = useSearch();
-  const [isLoading, setIsLoading] = useState(true);
-  const [videos, setVideos] = useState([]);
-
-  useEffect(() => {
-    if (searchTerm) {
-      getVideos(searchTerm).then((data) => {
-        const finalData = data.items.filter((item) => {
-          const { id } = item;
-          return id.kind !== 'youtube#channel';
-        });
-        setIsLoading(false);
-        setVideos(finalData);
-      });
-    }
-  }, [searchTerm]);
+  const { isLoading, videos } = useFetchVideos(searchTerm);
 
   return (
     <section className="homepage full-width">
